pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-graph-service"
version = "0.2.0"
description = "Python microservice for repository graph construction and context lookup"
readme = "README.md"
authors = [{ name = "Repo Graph Team" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "networkx>=3.3",
    "tree-sitter-languages>=1.9.0",
    "pygments>=2.18.0",
    "tqdm>=4.66.0",
    "grep-ast>=0.7.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.6.0",
    "SQLAlchemy>=2.0.37",
    "gitpython>=3.1.43",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
]

[tool.uvicorn]
app = "app.main:app"
host = "0.0.0.0"
port = 8000
reload = true
